Refactor OCR helpers into modules
