+++ server/routes/ocr.js
@@ -7,1010 +7,35 @@ import { admin } from '../firebase.js';
 import { db, ensureAppUserExists } from '../db.js';
 import { calculateMatch, extractCoffeeName } from '../utils/coffee.js';
 import { LOG_DIR } from '../utils/logging.js';
+import {
+  extractCoffeeAttributesFromText,
+  formatCoffeeAttributesSummary,
+  hasMeaningfulCoffeeData,
+  isMinimumCoffeeData,
+} from './ocr/coffee-attributes.js';
+import {
+  EVALUATION_RESPONSE_JSON_SCHEMA,
+  EVALUATION_RESPONSE_SCHEMA,
+  INSUFFICIENT_COFFEE_DATA_RESPONSE,
+  PROFILE_MISSING_RESPONSE,
+  applyLowDataAdjustments,
+  applyTasteMappingFallback,
+  buildDeterministicFallbackResponse,
+  buildLowConfidenceFallbackResponse,
+  isValidEvaluationResponse,
+  normalizeOpenAiJson,
+  resolveCorrectedText,
+} from './ocr/evaluation.js';
+import { extractStructuredMetadataFromText } from './ocr/structured-metadata.js';
+import {
+  formatTasteProfileSummary,
+  isTasteProfileComplete,
+  normalizeTasteProfileForEvaluation,
+} from './ocr/taste-profile.js';
 
 const router = express.Router();
 
 const GOOGLE_VISION_API_KEY = process.env.GOOGLE_VISION_API_KEY || ' ';
-// Schema-compliant fallback returned when the taste profile is missing or incomplete.
-const PROFILE_MISSING_RESPONSE = {
-  status: 'profile_missing',
-  verdict: null,
-  confidence: null,
-  verdict_explanation: {
-    user_preferences_summary:
-      'Tvoje preferencie: chuťový profil nie je dokončený, dáta chýbajú pre sladkosť, aciditu, horkosť a telo.',
-    coffee_profile_summary:
-      'Profil kávy: údaje o káve máme, ale bez preferencií nevieme zhrnúť chuťový profil pre sladkosť, aciditu, horkosť a telo.',
-    comparison_summary:
-      'Porovnanie s tvojím profilom: dáta chýbajú, doplň sladkosť, aciditu, horkosť, telo a 1–2 obľúbené tóny.',
-  },
-  insight: {
-    headline: 'Dokonči chuťový profil',
-    why: [
-      'Potrebujeme tvoje preferencie sladkosti, acidity, horkosti a tela.',
-      'Bez profilu by sme len hádali, čo ti chutí.',
-    ],
-    what_youll_like: [],
-    what_might_bother_you: [],
-    how_to_brew_for_better_match: [
-      'Vyplň krátky dotazník s chuťovými preferenciami.',
-    ],
-    recommended_alternatives: [
-      'Po dokončení profilu ti odporučíme vhodnejšie kávy.',
-    ],
-  },
-  disclaimer: 'Vyhodnotenie bude dostupné po dokončení chuťového profilu.',
-};
-
-const INSUFFICIENT_COFFEE_DATA_RESPONSE = {
-  status: 'insufficient_coffee_data',
-  verdict: null,
-  confidence: null,
-  verdict_explanation: {
-    user_preferences_summary:
-      'Tvoje preferencie: chuťové preferencie máme uložené, ale chýbajú detaily o káve.',
-    coffee_profile_summary:
-      'Profil kávy: dáta chýbajú pre chuťový profil (sladkosť, acidita, horkosť, telo a tóny), preto zostávame opatrní.',
-    comparison_summary:
-      'Porovnanie s tvojím profilom: dáta chýbajú, doplň pôvod, chuťové tóny, praženie alebo spracovanie (napr. rescan alebo cupping).',
-  },
-  insight: {
-    headline: 'Máme príliš málo údajov',
-    why: [
-      'Na presné hodnotenie potrebujeme viac údajov o káve.',
-      'Bez nich by sme len hádali zhodu.',
-    ],
-    what_youll_like: [],
-    what_might_bother_you: [],
-    how_to_brew_for_better_match: [
-      'Skús rescan balenia alebo urob malý test (cupping/jedna dávka) a doplň chuťové signály.',
-    ],
-    recommended_alternatives: [
-      'Zatiaľ sa pozri na kávy s jasne uvedeným pôvodom a chuťovými tónmi.',
-    ],
-  },
-  disclaimer: 'Vyhodnotenie bude možné po doplnení údajov o káve.',
-};
-
-const resolveTasteHintLabel = (value) => {
-  if (typeof value !== 'string') {
-    return null;
-  }
-  const normalized = value.toLowerCase().replace(/_/g, '-').trim();
-  if (normalized.includes('medium-dark') || normalized.includes('medium dark')) {
-    return 'medium-dark';
-  }
-  if (normalized.includes('dark')) {
-    return 'dark';
-  }
-  if (normalized.includes('light')) {
-    return 'light';
-  }
-  if (normalized.includes('medium')) {
-    return 'medium';
-  }
-  if (normalized.includes('washed') || normalized.includes('fully washed') || normalized.includes('wet')) {
-    return 'washed';
-  }
-  if (normalized.includes('natural') || normalized.includes('dry')) {
-    return 'natural';
-  }
-  return null;
-};
-
-const buildTasteMappingSentences = ({ roastLevel, processing } = {}) => {
-  const sentences = [];
-  const roastLabel = resolveTasteHintLabel(roastLevel);
-  const processingLabel = resolveTasteHintLabel(processing);
-
-  if (roastLabel === 'dark' || roastLabel === 'medium-dark') {
-    sentences.push('Tmavé alebo stredne tmavé praženie naznačuje vyššiu horkosť a nižšiu aciditu.');
-  }
-  if (roastLabel === 'light') {
-    sentences.push('Svetlé praženie zvyčajne prináša vyššiu aciditu a ovocnejšie tóny.');
-  }
-  if (processingLabel === 'natural') {
-    sentences.push('Natural spracovanie zvýrazňuje ovocnosť a sladkosť.');
-  }
-  if (processingLabel === 'washed') {
-    sentences.push('Washed spracovanie prináša čistotu a vyššiu aciditu.');
-  }
-
-  return sentences;
-};
-
-const summaryHasKeywords = (summary, keywords) =>
-  keywords.every((keyword) => summary.includes(keyword));
-
-const addTasteMappingFallbackToSummary = (summary, coffeeAttributes) => {
-  if (typeof summary !== 'string') {
-    return summary;
-  }
-  const roastLevel =
-    coffeeAttributes?.roast_level ?? coffeeAttributes?.roastLevel ?? coffeeAttributes?.structured_metadata?.roast_level;
-  const processing = coffeeAttributes?.processing ?? coffeeAttributes?.structured_metadata?.processing;
-  const mappingSentences = buildTasteMappingSentences({ roastLevel, processing });
-  if (mappingSentences.length === 0) {
-    return summary;
-  }
-
-  const loweredSummary = summary.toLowerCase();
-  const keywordSets = [
-    { sentence: mappingSentences.find((item) => item.includes('horkosť')) || null, keywords: ['horkosť', 'acidita'] },
-    { sentence: mappingSentences.find((item) => item.includes('Svetlé')) || null, keywords: ['acidita', 'ovoc'] },
-    { sentence: mappingSentences.find((item) => item.includes('Natural')) || null, keywords: ['ovoc', 'sladk'] },
-    { sentence: mappingSentences.find((item) => item.includes('Washed')) || null, keywords: ['čist', 'acidita'] },
-  ];
-
-  const sentencesToAppend = keywordSets
-    .filter(({ sentence }) => sentence)
-    .filter(({ keywords }) => !summaryHasKeywords(loweredSummary, keywords))
-    .map(({ sentence }) => sentence);
-
-  if (sentencesToAppend.length === 0) {
-    return summary;
-  }
-
-  return `${summary} ${sentencesToAppend.join(' ')}`.trim();
-};
-
-const buildDeterministicFallbackResponse = ({ preferences, coffeeAttributes, correctedText }) => {
-  const structured =
-    coffeeAttributes && typeof coffeeAttributes === 'object'
-      ? coffeeAttributes.structured_metadata || {}
-      : {};
-  const safePreferences = preferences || {};
-  const matchScore = calculateMatch(
-    correctedText || '',
-    safePreferences,
-    structured
-  );
-  const normalizedScore =
-    typeof matchScore === 'number' && Number.isFinite(matchScore) ? matchScore : 50;
-  const verdict = normalizedScore >= 75 ? 'suitable' : 'not_suitable';
-  const confidence = Number((normalizedScore / 100).toFixed(2));
-  const origin = coffeeAttributes?.origin ?? structured?.origin;
-  const roastLevel =
-    coffeeAttributes?.roast_level ?? coffeeAttributes?.roastLevel ?? structured?.roast_level;
-  const flavorNotes =
-    coffeeAttributes?.flavor_notes ??
-    coffeeAttributes?.flavorNotes ??
-    structured?.flavor_notes ??
-    structured?.flavorNotes;
-  const processing = coffeeAttributes?.processing ?? structured?.processing;
-  const varietals = coffeeAttributes?.varietals ?? structured?.varietals;
-
-  const profileBits = [
-    origin ? `pôvod: ${origin}` : null,
-    roastLevel ? `praženie: ${roastLevel}` : null,
-    processing ? `spracovanie: ${processing}` : null,
-    Array.isArray(flavorNotes) && flavorNotes.length > 0
-      ? `tóny: ${flavorNotes.join(', ')}`
-      : null,
-    Array.isArray(varietals) && varietals.length > 0
-      ? `odrody: ${varietals.join(', ')}`
-      : null,
-  ].filter(Boolean);
-
-  const baseCoffeeProfileSummary =
-    profileBits.length > 0
-      ? `Profil kávy: obsahuje ${profileBits.join(', ')}.`
-      : 'Profil kávy: máme len základné údaje z OCR textu a balenia.';
-  const coffeeProfileSummary = addTasteMappingFallbackToSummary(
-    baseCoffeeProfileSummary,
-    coffeeAttributes
-  );
-
-  const formatPreference = (value) =>
-    value === null || value === undefined || value === '' ? 'neznáme' : value;
-  const extraPreferenceBits = [];
-  if (safePreferences.intensity !== undefined && safePreferences.intensity !== null) {
-    extraPreferenceBits.push(`intenzita ${formatPreference(safePreferences.intensity)}`);
-  }
-  if (
-    safePreferences.experimentalism !== undefined &&
-    safePreferences.experimentalism !== null
-  ) {
-    extraPreferenceBits.push(
-      `experimentalnosť ${formatPreference(safePreferences.experimentalism)}`
-    );
-  }
-  const extraPreferenceSummary =
-    extraPreferenceBits.length > 0 ? `, ${extraPreferenceBits.join(', ')}` : '';
-  const userPreferencesSummary = `Tvoje preferencie: sladkosť ${formatPreference(
-    safePreferences.sweetness
-  )}, acidita ${formatPreference(safePreferences.acidity)}, horkosť ${formatPreference(
-    safePreferences.bitterness
-  )}, telo ${formatPreference(safePreferences.body)}${extraPreferenceSummary}.`;
-  const extraComparisonSummary =
-    extraPreferenceBits.length > 0
-      ? ` Zohľadnili sme aj ${extraPreferenceBits.join(' a ')}.`
-      : '';
-  const comparisonSummary =
-    verdict === 'suitable'
-      ? `Porovnanie s tvojím profilom: textová zhoda vychádza na ${Math.round(
-          normalizedScore
-        )} %, preto kávu hodnotíme ako vhodnú.${extraComparisonSummary}`
-      : `Porovnanie s tvojím profilom: textová zhoda vychádza na ${Math.round(
-          normalizedScore
-        )} %, preto kávu hodnotíme ako menej vhodnú.${extraComparisonSummary}`;
-
-  return {
-    status: 'ok',
-    verdict,
-    confidence,
-    verdict_explanation: {
-      user_preferences_summary: userPreferencesSummary,
-      coffee_profile_summary: coffeeProfileSummary,
-      comparison_summary: comparisonSummary,
-    },
-    insight: {
-      headline: verdict === 'suitable' ? 'Vyzerá to vhodne' : 'Zváž inú voľbu',
-      why:
-        verdict === 'suitable'
-          ? [
-              'Zistené údaje zodpovedajú preferovanému chuťovému profilu.',
-              'Celkové skóre zhody je nad odporúčaným prahom.',
-            ]
-          : [
-              'Zistené údaje sa zhodujú len čiastočne s preferenciami.',
-              'Celkové skóre je pod odporúčaným prahom.',
-            ],
-      what_youll_like: verdict === 'suitable' ? ['Stabilný profil podľa tvojich preferencií.'] : [],
-      what_might_bother_you:
-        verdict === 'suitable' ? [] : ['Profil môže byť odlišný od tvojich preferencií.'],
-      how_to_brew_for_better_match: [
-        'Ak chceš vyššiu zhodu, dolaď mletie a dávku podľa chuti.',
-      ],
-      recommended_alternatives:
-        verdict === 'suitable' ? [] : ['Pozri sa na kávy s podobnými chuťovými tónmi.'],
-    },
-    disclaimer: 'Výsledok je založený na dostupných údajoch z OCR.',
-  };
-};
-
-const EVALUATION_RESPONSE_SCHEMA = `JSON schema (strict):
-{
-  "status": "ok" | "profile_missing" | "insufficient_coffee_data",
-  "verdict": "suitable" | "not_suitable" | "uncertain" | null,
-  "confidence": number | null,
-  "verdict_explanation": {
-    "user_preferences_summary": string,
-    "coffee_profile_summary": string,
-    "comparison_summary": string
-  },
-  "insight": {
-    "headline": string,
-    "why": string[],
-    "what_youll_like": string[],
-    "what_might_bother_you": string[],
-    "how_to_brew_for_better_match": string[],
-    "recommended_alternatives": string[]
-  },
-  "disclaimer": string
-}
-
-Return JSON only. Do not include markdown fences or extra text.`;
-
-const EVALUATION_RESPONSE_JSON_SCHEMA = {
-  type: 'object',
-  additionalProperties: false,
-  required: [
-    'status',
-    'verdict',
-    'confidence',
-    'verdict_explanation',
-    'insight',
-    'disclaimer',
-  ],
-  properties: {
-    status: {
-      type: 'string',
-      enum: ['ok', 'profile_missing', 'insufficient_coffee_data'],
-    },
-    verdict: {
-      type: ['string', 'null'],
-      enum: ['suitable', 'not_suitable', 'uncertain', null],
-    },
-    confidence: {
-      type: ['number', 'null'],
-    },
-    verdict_explanation: {
-      type: 'object',
-      additionalProperties: false,
-      required: [
-        'user_preferences_summary',
-        'coffee_profile_summary',
-        'comparison_summary',
-      ],
-      properties: {
-        user_preferences_summary: { type: 'string' },
-        coffee_profile_summary: { type: 'string' },
-        comparison_summary: { type: 'string' },
-      },
-    },
-    insight: {
-      type: 'object',
-      additionalProperties: false,
-      required: [
-        'headline',
-        'why',
-        'what_youll_like',
-        'what_might_bother_you',
-        'how_to_brew_for_better_match',
-        'recommended_alternatives',
-      ],
-      properties: {
-        headline: { type: 'string' },
-        why: { type: 'array', items: { type: 'string' } },
-        what_youll_like: { type: 'array', items: { type: 'string' } },
-        what_might_bother_you: { type: 'array', items: { type: 'string' } },
-        how_to_brew_for_better_match: { type: 'array', items: { type: 'string' } },
-        recommended_alternatives: { type: 'array', items: { type: 'string' } },
-      },
-    },
-    disclaimer: {
-      type: 'string',
-    },
-  },
-};
-
-const normalizeOpenAiJson = (value) => {
-  if (!value) {
-    return value;
-  }
-  return value.replace(/```json\s*/i, '').replace(/```$/i, '').trim();
-};
-
-const resolveCorrectedText = ({ corrected_text, coffee_attributes } = {}) => {
-  const normalizeText = (value) =>
-    typeof value === 'string' && value.trim().length > 0 ? value.trim() : null;
-  const topLevelText = normalizeText(corrected_text);
-  if (topLevelText) {
-    return topLevelText;
-  }
-  return normalizeText(coffee_attributes?.corrected_text);
-};
-
-const BRANDED_COFFEE_ALLOWLIST = ['Lavazza', 'Illy', 'Segafredo', 'Kimbo', 'Pellini', 'Bazzara'];
-
-const normalizeCoffeeValue = (value) => {
-  if (typeof value !== 'string') {
-    return null;
-  }
-  const cleaned = value.replace(/\s+/g, ' ').trim();
-  return cleaned.length > 0 ? cleaned : null;
-};
-
-const parseDelimitedList = (value) => {
-  if (!value || typeof value !== 'string') {
-    return null;
-  }
-  const items = value
-    .split(/[,;|/]+/)
-    .map((item) => item.replace(/[()\[\]]/g, '').trim())
-    .filter(Boolean);
-  return items.length > 0 ? items : null;
-};
-
-const resolveTasteVector = (profile) => {
-  if (!profile || typeof profile !== 'object') {
-    return null;
-  }
-  if (profile.taste_vector && typeof profile.taste_vector === 'object') {
-    return profile.taste_vector;
-  }
-  if (profile.preferences && typeof profile.preferences === 'object') {
-    return profile.preferences;
-  }
-  return profile;
-};
-
-const formatTasteProfileSummary = (profile) => {
-  const vector = resolveTasteVector(profile);
-  if (!vector || typeof vector !== 'object') {
-    return null;
-  }
-  const { sweetness, acidity, bitterness, body, intensity, experimentalism } = vector;
-  const values = [sweetness, acidity, bitterness, body, intensity, experimentalism];
-  if (!values.some((value) => typeof value === 'number' && Number.isFinite(value))) {
-    return null;
-  }
-  const formatValue = (value) =>
-    typeof value === 'number' && Number.isFinite(value) ? `${value}/10` : 'neznáme';
-  const summaryBits = [
-    `sladkosť ${formatValue(sweetness)}`,
-    `acidita ${formatValue(acidity)}`,
-    `horkosť ${formatValue(bitterness)}`,
-    `telo ${formatValue(body)}`,
-  ];
-  if (intensity !== undefined && intensity !== null) {
-    summaryBits.push(`intenzita ${formatValue(intensity)}`);
-  }
-  if (experimentalism !== undefined && experimentalism !== null) {
-    summaryBits.push(`experimentalnosť ${formatValue(experimentalism)}`);
-  }
-  return summaryBits.join(', ');
-};
-
-const formatCoffeeAttributesSummary = (attributes) => {
-  if (!attributes || typeof attributes !== 'object') {
-    return null;
-  }
-  const structured =
-    attributes.structured_metadata && typeof attributes.structured_metadata === 'object'
-      ? attributes.structured_metadata
-      : {};
-
-  const getValue = (record, keys) =>
-    keys.reduce((acc, key) => (acc !== undefined ? acc : record?.[key]), undefined);
-
-  const origin = getValue(attributes, ['origin']) ?? getValue(structured, ['origin']);
-  const roastLevel =
-    getValue(attributes, ['roast_level', 'roastLevel']) ??
-    getValue(structured, ['roast_level', 'roastLevel']);
-  const processing =
-    getValue(attributes, ['processing']) ?? getValue(structured, ['processing']);
-  const roaster =
-    getValue(attributes, ['roaster', 'roastery', 'brand']) ??
-    getValue(structured, ['roaster', 'roastery', 'brand']);
-  const flavorNotes =
-    getValue(attributes, ['flavor_notes', 'flavorNotes']) ??
-    getValue(structured, ['flavor_notes', 'flavorNotes']);
-  const varietals =
-    getValue(attributes, ['varietals']) ?? getValue(structured, ['varietals']);
-
-  const summaryBits = [
-    origin ? `pôvod ${origin}` : null,
-    roastLevel ? `praženie ${roastLevel}` : null,
-    processing ? `spracovanie ${processing}` : null,
-    roaster ? `pražiareň ${roaster}` : null,
-    Array.isArray(flavorNotes) && flavorNotes.length > 0
-      ? `tóny ${flavorNotes.join(', ')}`
-      : typeof flavorNotes === 'string' && flavorNotes.trim().length > 0
-      ? `tóny ${flavorNotes}`
-      : null,
-    Array.isArray(varietals) && varietals.length > 0
-      ? `odrody ${varietals.join(', ')}`
-      : typeof varietals === 'string' && varietals.trim().length > 0
-      ? `odrody ${varietals}`
-      : null,
-  ].filter(Boolean);
-
-  return summaryBits.length > 0 ? summaryBits.join(', ') : null;
-};
-
-const extractCoffeeAttributesFromText = (text) => {
-  if (!text || typeof text !== 'string') {
-    return {
-      origin: null,
-      roast_level: null,
-      flavor_notes: null,
-      processing: null,
-      varietals: null,
-      brand: null,
-      keywords: null,
-    };
-  }
-
-  const normalizedText = text.replace(/\r/g, '');
-  const lowerText = normalizedText.toLowerCase();
-
-  const extractLabelValue = (labels) => {
-    for (const label of labels) {
-      const regex = new RegExp(`${label}\\s*[:\\-]\\s*([^\\n]+)`, 'i');
-      const match = normalizedText.match(regex);
-      if (match && match[1]) {
-        return normalizeCoffeeValue(match[1]);
-      }
-    }
-    return null;
-  };
-
-  const origin = extractLabelValue(['origin', 'pôvod', 'povod', 'country', 'region']) || null;
-  const flavorNotesRaw = extractLabelValue([
-    'flavor notes',
-    'flavour notes',
-    'tóny',
-    'chut',
-    'notes',
-  ]);
-  const varietalsRaw = extractLabelValue([
-    'varietal',
-    'varietals',
-    'variety',
-    'odroda',
-    'odrody',
-  ]);
-
-  const processingKeywords = [
-    { label: 'washed', keywords: ['washed', 'fully washed', 'wet process', 'mokre'] },
-    { label: 'natural', keywords: ['natural', 'dry process', 'sušené', 'susene'] },
-    { label: 'honey', keywords: ['honey', 'honey process', 'pulped natural'] },
-    { label: 'anaerobic', keywords: ['anaerobic', 'anaeróbne', 'anaerobne'] },
-    { label: 'semi-washed', keywords: ['semi-washed', 'semi washed', 'wet hulled'] },
-    { label: 'carbonic maceration', keywords: ['carbonic', 'maceration'] },
-  ];
-  const processingKeywordMatch = processingKeywords.find(({ keywords }) =>
-    keywords.some((keyword) => lowerText.includes(keyword))
-  );
-  const processing =
-    extractLabelValue(['processing', 'process', 'spracovanie']) ||
-    processingKeywordMatch?.label ||
-    null;
-
-  const roastLevelRaw = extractLabelValue(['roast', 'praženie', 'prazenie']);
-  const roastLevelKeywords = [
-    { label: 'light', keywords: ['light', 'svetla', 'cinnamon', 'blonde'] },
-    { label: 'medium', keywords: ['medium', 'stredna', 'city'] },
-    { label: 'medium-dark', keywords: ['medium-dark', 'medium dark', 'full city'] },
-    { label: 'dark', keywords: ['dark', 'tmava', 'french', 'italian', 'espresso'] },
-  ];
-  const roastKeywordMatch = roastLevelKeywords.find(({ keywords }) =>
-    keywords.some((keyword) => lowerText.includes(keyword))
-  );
-  const roastLevel =
-    normalizeCoffeeValue(roastLevelRaw) ||
-    roastKeywordMatch?.label ||
-    null;
-
-  const brandRaw = extractLabelValue([
-    'brand',
-    'roaster',
-    'roastery',
-    'roaster name',
-    'pražiareň',
-    'praziaren',
-  ]);
-  const brandAllowlistMatch = BRANDED_COFFEE_ALLOWLIST.find((brandName) =>
-    lowerText.includes(brandName.toLowerCase())
-  );
-  const brand = normalizeCoffeeValue(brandRaw || brandAllowlistMatch);
-
-  const keywordMatches = [
-    roastKeywordMatch?.label,
-    processingKeywordMatch?.label,
-    brandAllowlistMatch ? `brand:${brandAllowlistMatch}` : null,
-    lowerText.includes('single origin') ? 'single origin' : null,
-    lowerText.includes('blend') ? 'blend' : null,
-    lowerText.includes('arabica') ? 'arabica' : null,
-    lowerText.includes('robusta') ? 'robusta' : null,
-    lowerText.includes('espresso') ? 'espresso' : null,
-    lowerText.includes('filter') ? 'filter' : null,
-  ].filter(Boolean);
-
-  return {
-    origin,
-    roast_level: roastLevel,
-    flavor_notes: flavorNotesRaw ? parseDelimitedList(flavorNotesRaw) : null,
-    processing: normalizeCoffeeValue(processing),
-    varietals: varietalsRaw ? parseDelimitedList(varietalsRaw) : null,
-    brand,
-    keywords: keywordMatches.length > 0 ? keywordMatches : null,
-  };
-};
-
-const hasMeaningfulCoffeeData = (coffeeAttributes) => {
-  if (!coffeeAttributes || typeof coffeeAttributes !== 'object') {
-    return false;
-  }
-
-  const { structured_metadata } = coffeeAttributes;
-  const structured =
-    structured_metadata && typeof structured_metadata === 'object' ? structured_metadata : {};
-
-  const getValue = (record, keys) =>
-    keys.reduce((acc, key) => (acc !== undefined ? acc : record?.[key]), undefined);
-  const hasValue = (value) => {
-    if (Array.isArray(value)) {
-      return value.length > 0;
-    }
-    if (typeof value === 'string') {
-      return value.trim().length > 0;
-    }
-    return value !== null && value !== undefined;
-  };
-
-  const origin = getValue(coffeeAttributes, ['origin']) ?? getValue(structured, ['origin']);
-  const roastLevel =
-    getValue(coffeeAttributes, ['roast_level', 'roastLevel']) ??
-    getValue(structured, ['roast_level', 'roastLevel']);
-  const flavorNotes =
-    getValue(coffeeAttributes, ['flavor_notes', 'flavorNotes']) ??
-    getValue(structured, ['flavor_notes', 'flavorNotes']);
-  const processing =
-    getValue(coffeeAttributes, ['processing']) ?? getValue(structured, ['processing']);
-  const varietals =
-    getValue(coffeeAttributes, ['varietals']) ?? getValue(structured, ['varietals']);
-  const brand =
-    getValue(coffeeAttributes, ['brand', 'roaster', 'roastery', 'roaster_name']) ??
-    getValue(structured, ['brand', 'roaster', 'roastery', 'roaster_name']);
-  const keywords =
-    getValue(coffeeAttributes, ['keywords']) ?? getValue(structured, ['keywords']);
-
-  const hasCoreProfileDetails = [origin, flavorNotes, varietals].some(hasValue);
-  const hasRoastOrProcessing = [roastLevel, processing].some(hasValue);
-  const brandLabel = typeof brand === 'string' ? brand.trim().toLowerCase() : '';
-  const isAllowlistedBrand = BRANDED_COFFEE_ALLOWLIST.some((allowed) =>
-    brandLabel.includes(allowed.toLowerCase())
-  );
-  const hasKeywordSignals = Array.isArray(keywords) && keywords.length > 0;
-  const hasBrandSignal = hasValue(brand);
-
-  // Require structured attributes before AI evaluation.
-  // Allow recognized brands with roast/processing data to pass the check.
-  return (
-    hasCoreProfileDetails ||
-    hasRoastOrProcessing ||
-    hasKeywordSignals ||
-    (isAllowlistedBrand && hasRoastOrProcessing) ||
-    hasBrandSignal
-  );
-};
-
-const isMinimumCoffeeData = (coffeeAttributes) => {
-  if (!coffeeAttributes || typeof coffeeAttributes !== 'object') {
-    return false;
-  }
-
-  const { structured_metadata } = coffeeAttributes;
-  const structured =
-    structured_metadata && typeof structured_metadata === 'object' ? structured_metadata : {};
-
-  const getValue = (record, keys) =>
-    keys.reduce((acc, key) => (acc !== undefined ? acc : record?.[key]), undefined);
-  const hasValue = (value) => {
-    if (Array.isArray(value)) {
-      return value.length > 0;
-    }
-    if (typeof value === 'string') {
-      return value.trim().length > 0;
-    }
-    return value !== null && value !== undefined;
-  };
-
-  const origin = getValue(coffeeAttributes, ['origin']) ?? getValue(structured, ['origin']);
-  const roastLevel =
-    getValue(coffeeAttributes, ['roast_level', 'roastLevel']) ??
-    getValue(structured, ['roast_level', 'roastLevel']);
-  const flavorNotes =
-    getValue(coffeeAttributes, ['flavor_notes', 'flavorNotes']) ??
-    getValue(structured, ['flavor_notes', 'flavorNotes']);
-  const processing =
-    getValue(coffeeAttributes, ['processing']) ?? getValue(structured, ['processing']);
-  const varietals =
-    getValue(coffeeAttributes, ['varietals']) ?? getValue(structured, ['varietals']);
-
-  const hasCoreProfileDetails = [origin, flavorNotes, varietals].some(hasValue);
-  const hasRoastOrProcessing = [roastLevel, processing].some(hasValue);
-
-  return hasRoastOrProcessing && !hasCoreProfileDetails;
-};
-
-const buildLowConfidenceFallbackResponse = ({ preferences, coffeeAttributes, correctedText }) => {
-  const base = buildDeterministicFallbackResponse({
-    preferences,
-    coffeeAttributes,
-    correctedText,
-  });
-  const loweredConfidence =
-    typeof base.confidence === 'number' && Number.isFinite(base.confidence)
-      ? Math.min(Number((base.confidence * 0.6).toFixed(2)), 0.45)
-      : 0.3;
-
-  return {
-    ...base,
-    confidence: loweredConfidence,
-    verdict: base.verdict ?? 'uncertain',
-    verdict_explanation: {
-      ...base.verdict_explanation,
-      coffee_profile_summary: `${base.verdict_explanation.coffee_profile_summary} Dáta sú slabšie, takže hodnotenie je orientačné.`,
-    },
-    insight: {
-      ...base.insight,
-      why: [
-        ...base.insight.why,
-        'Použité boli iba čiastočné signály z textu.',
-      ],
-    },
-    disclaimer: 'Výsledok je orientačný a založený na obmedzených údajoch z OCR.',
-  };
-};
-
-const applyTasteMappingFallback = (response, coffeeAttributes) => {
-  if (!response || typeof response !== 'object') {
-    return response;
-  }
-  const verdictExplanation = response.verdict_explanation;
-  if (!verdictExplanation || typeof verdictExplanation !== 'object') {
-    return response;
-  }
-  const coffeeProfileSummary = verdictExplanation.coffee_profile_summary;
-  const enrichedSummary = addTasteMappingFallbackToSummary(
-    coffeeProfileSummary,
-    coffeeAttributes
-  );
-  if (enrichedSummary === coffeeProfileSummary) {
-    return response;
-  }
-  return {
-    ...response,
-    verdict_explanation: {
-      ...verdictExplanation,
-      coffee_profile_summary: enrichedSummary,
-    },
-  };
-};
-
-const applyLowDataAdjustments = (response) => {
-  if (!response || typeof response !== 'object') {
-    return response;
-  }
-  const loweredConfidence =
-    typeof response.confidence === 'number' && Number.isFinite(response.confidence)
-      ? Math.min(Number((response.confidence * 0.6).toFixed(2)), 0.45)
-      : 0.3;
-  const disclaimer = response.disclaimer || '';
-  const lowDataSentence =
-    'Máme len minimum údajov (napr. iba praženie alebo spracovanie), preto je hodnotenie orientačné.';
-  const nextDisclaimer = disclaimer.includes('minimum údajov')
-    ? disclaimer
-    : `${disclaimer}${disclaimer ? ' ' : ''}${lowDataSentence} Skús prosím rescan etikety.`;
-
-  return {
-    ...response,
-    confidence: loweredConfidence,
-    verdict_explanation: response.verdict_explanation ?? {},
-    disclaimer: nextDisclaimer,
-  };
-};
-
-const isValidEvaluationResponse = (value) => {
-  if (!value || typeof value !== 'object' || Array.isArray(value)) {
-    return false;
-  }
-
-  const requiredKeys = [
-    'status',
-    'verdict',
-    'confidence',
-    'verdict_explanation',
-    'insight',
-    'disclaimer',
-  ];
-  if (!requiredKeys.every((key) => Object.prototype.hasOwnProperty.call(value, key))) {
-    return false;
-  }
-
-  const {
-    status,
-    verdict,
-    confidence,
-    verdict_explanation,
-    insight,
-    disclaimer,
-  } = value;
-
-  if (
-    status !== 'ok' &&
-    status !== 'profile_missing' &&
-    status !== 'insufficient_coffee_data'
-  ) {
-    return false;
-  }
-
-  const validVerdicts = ['suitable', 'not_suitable', 'uncertain'];
-  if (verdict !== null && !validVerdicts.includes(verdict)) {
-    return false;
-  }
-
-  if (confidence !== null && (typeof confidence !== 'number' || Number.isNaN(confidence))) {
-    return false;
-  }
-
-  if (!verdict_explanation || typeof verdict_explanation !== 'object') {
-    return false;
-  }
-
-  if (
-    typeof verdict_explanation.user_preferences_summary !== 'string' ||
-    typeof verdict_explanation.coffee_profile_summary !== 'string' ||
-    typeof verdict_explanation.comparison_summary !== 'string'
-  ) {
-    return false;
-  }
-
-  if (!insight || typeof insight !== 'object') {
-    return false;
-  }
-
-  const isStringArray = (value) =>
-    Array.isArray(value) && value.every((item) => typeof item === 'string');
-
-  if (
-    typeof insight.headline !== 'string' ||
-    !isStringArray(insight.why) ||
-    !isStringArray(insight.what_youll_like) ||
-    !isStringArray(insight.what_might_bother_you) ||
-    !isStringArray(insight.how_to_brew_for_better_match) ||
-    !isStringArray(insight.recommended_alternatives)
-  ) {
-    return false;
-  }
-
-  if (typeof disclaimer !== 'string') {
-    return false;
-  }
-
-  if (status === 'ok') {
-    if (verdict === null || confidence === null) {
-      return false;
-    }
-  } else {
-    if (verdict !== null || confidence !== null) {
-      return false;
-    }
-  }
-
-  return true;
-};
-
-// OCR evaluation uses the 4D core; extra taste dimensions are accepted but ignored here.
-const isTasteProfileComplete = (profile) => {
-  if (!profile || typeof profile !== 'object') {
-    return false;
-  }
-
-  if (profile.is_complete === true || profile.taste_profile_completed === true) {
-    return true;
-  }
-
-  const tasteVector =
-    profile.taste_vector && typeof profile.taste_vector === 'object'
-      ? profile.taste_vector
-      : profile.preferences && typeof profile.preferences === 'object'
-      ? profile.preferences
-      : profile;
-
-  const values = [
-    tasteVector.sweetness,
-    tasteVector.acidity,
-    tasteVector.bitterness,
-    tasteVector.body,
-  ];
-
-  return values.every(
-    (value) => typeof value === 'number' && Number.isFinite(value) && value >= 0 && value <= 10
-  );
-};
-
-const normalizeTasteProfileForEvaluation = (profile) => {
-  if (!profile || typeof profile !== 'object') {
-    return profile;
-  }
-
-  const tasteVector =
-    profile.taste_vector && typeof profile.taste_vector === 'object'
-      ? profile.taste_vector
-      : profile.preferences && typeof profile.preferences === 'object'
-      ? profile.preferences
-      : null;
-
-  return tasteVector ? { ...profile, ...tasteVector } : profile;
-};
-
-const parseStructuredResponseJson = (content) => {
-  if (!content || typeof content !== 'string') {
-    return null;
-  }
-  const trimmed = content.trim();
-  if (!trimmed) {
-    return null;
-  }
-  try {
-    return JSON.parse(trimmed);
-  } catch (error) {
-    const match = trimmed.match(/\{[\s\S]*\}/);
-    if (!match) {
-      return null;
-    }
-    try {
-      return JSON.parse(match[0]);
-    } catch (innerError) {
-      return null;
-    }
-  }
-};
-
-const extractStructuredMetadataFromText = async (text) => {
-  if (!text || typeof text !== 'string' || text.trim().length === 0) {
-    return { structured_metadata: null, structured_confidence: null };
-  }
-
-  if (!process.env.OPENAI_API_KEY) {
-    return { structured_metadata: null, structured_confidence: null };
-  }
-
-  const prompt = `Z OCR textu extrahuj štruktúrované údaje o káve.
-Vráť JSON s dvomi kľúčmi: structured_metadata a structured_confidence.
-structured_metadata musí obsahovať presne tieto kľúče:
-name, roast_level, processing, origin, varietals, flavor_notes, roaster, roast_date, coffee_type, weight, brew_methods.
-Použi null, ak hodnota nie je z textu jasná. Odrody (varietals), flavor_notes a brew_methods vráť ako pole stringov.
-structured_confidence má rovnaké kľúče a obsahuje číslo 0 až 1 alebo null podľa istoty.
-Nepridávaj ďalší text, iba čistý JSON.
-
-OCR text:
-${text}`;
-
-  try {
-    console.log('📤 [OpenAI] Structured metadata prompt meta:', {
-      model: 'gpt-4o',
-      chars: prompt.length,
-    });
-    const response = await axios.post(
-      'https://api.openai.com/v1/chat/completions',
-      {
-        model: 'gpt-4o',
-        messages: [
-          {
-            role: 'system',
-            content:
-              'Si expert na kávu. Extrahuješ štruktúrované údaje z OCR textu etikiet kávy.',
-          },
-          { role: 'user', content: prompt },
-        ],
-        temperature: 0.2,
-      },
-      {
-        headers: {
-          Authorization: `Bearer ${process.env.OPENAI_API_KEY}`,
-          'Content-Type': 'application/json',
-        },
-      }
-    );
-
-    console.log('📥 [OpenAI] Structured metadata response meta:', {
-      id: response.data?.id,
-      model: response.data?.model,
-      usage: response.data?.usage,
-    });
-
-    const content = response.data?.choices?.[0]?.message?.content || '';
-    const parsed = parseStructuredResponseJson(content);
-    if (!parsed || typeof parsed !== 'object') {
-      return { structured_metadata: null, structured_confidence: null };
-    }
-
-    const structuredMetadata =
-      parsed.structured_metadata ||
-      parsed.structuredMetadata ||
-      parsed.metadata ||
-      parsed.data ||
-      null;
-    const structuredConfidence =
-      parsed.structured_confidence ||
-      parsed.structuredConfidence ||
-      parsed.confidence ||
-      null;
-
-    const resolvedMetadata =
-      structuredMetadata && typeof structuredMetadata === 'object'
-        ? structuredMetadata
-        : parsed;
-    const resolvedConfidence =
-      structuredConfidence && typeof structuredConfidence === 'object'
-        ? structuredConfidence
-        : null;
-
-    return {
-      structured_metadata: resolvedMetadata,
-      structured_confidence: resolvedConfidence,
-    };
-  } catch (error) {
-    console.error('Structured metadata AI error:', error?.message ?? error);
-    return { structured_metadata: null, structured_confidence: null };
-  }
-};
 
 // ========== OCR ENDPOINTS ==========
